Bugfix: anchored links on the homepage
This fixes a bug caused by the [anchored headings](https://github.com/alphagov/tech-docs-gem/blob/master/lib/assets/javascripts/_modules/anchored-headings.js) feature inherited from the `govuk_tech_docs` gem. This javascript injects an anchor link for each heading which becomes visible on hover, displaying as a link-in-chain icon.

Unfortunately, this link expects there to be an `id` attribute set on each heading, and otherwise links to `#undefined`, which doesn't work.

This fixes the links by adding the required id attributes.

It's still a slightly odd behaviour on the homepage, where these headings are already links to other pages, and so the anchor link is probably not that useful and possibly confusing.  Unfortunately, the javascript doesn't have any ability to opt-out of this behaviour, other than by using another element such as a `<div>` instead of the  `<h2>`. We could possibly override it with some slightly hacky additional javascript or CSS, but that doesn't seem ideal.

So in the short term perhaps we should keep this feature and just fix it so that the anchor links work?